// src/domain/models/job.ts
export type JobStatus = "pending" | "processing" | "completed" | "failed";

export type CreateSumJobDTO = {
  id?: string;
  status?: JobStatus;
  input: {
    num1: number;
    num2: number;
  };
  result?: {
    sum: number;
  };
  error?: string;
  createdAt?: Date;
  updatedAt?: Date;
};

export type ISumJobToJSON = {
  id: string;
  status: JobStatus;
  input: {
    num1: number;
    num2: number;
  };
  result?: {
    sum: number;
  };
  error?: string;
  createdAt: Date;
  updatedAt: Date;
};

export interface ISumJob {
  toJSON(): ISumJobToJSON;
}